Rename launcher module to zapscript
