+++ pkg/zapscript/commands.go
@@ -18,7 +18,7 @@ You should have received a copy of the GNU General Public License
 along with Zaparoo Core.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-package launcher
+package zapscript
 
 import (
 	"errors"
