+++ pkg/zapscript/launch.go
@@ -1,4 +1,4 @@
-package launcher
+package zapscript
 
 import (
 	"fmt"
